Relay local/remote multiaddress retrieval to Transport implementations to get rid of subclass checks
