+++ libp2p/src/main/kotlin/io/libp2p/transport/implementation/ConnectionBuilder.kt
@@ -5,7 +5,6 @@ import io.libp2p.core.Connection
 import io.libp2p.core.ConnectionHandler
 import io.libp2p.core.P2PChannel
 import io.libp2p.core.PeerId
-import io.libp2p.core.transport.Transport
 import io.libp2p.etc.REMOTE_PEER_ID
 import io.libp2p.etc.types.forward
 import io.libp2p.transport.ConnectionUpgrader
@@ -14,7 +13,7 @@ import io.netty.channel.ChannelInitializer
 import java.util.concurrent.CompletableFuture
 
 class ConnectionBuilder(
-    private val transport: Transport,
+    private val transport: NettyTransport,
     private val upgrader: ConnectionUpgrader,
     private val connHandler: ConnectionHandler,
     private val initiator: Boolean,
