+++ libp2p/src/main/kotlin/io/libp2p/transport/tcp/TcpTransport.kt
@@ -1,20 +1,17 @@
 package io.libp2p.transport.tcp
 
-import io.libp2p.core.InternalErrorException
 import io.libp2p.core.multiformats.Multiaddr
 import io.libp2p.core.multiformats.Protocol.DNSADDR
-import io.libp2p.core.multiformats.Protocol.IP4
-import io.libp2p.core.multiformats.Protocol.IP6
 import io.libp2p.core.multiformats.Protocol.P2PCIRCUIT
 import io.libp2p.core.multiformats.Protocol.TCP
 import io.libp2p.core.multiformats.Protocol.WS
+import io.libp2p.etc.util.MultiaddrUtils
 import io.libp2p.transport.ConnectionUpgrader
 import io.libp2p.transport.implementation.ConnectionBuilder
-import io.libp2p.transport.implementation.NettyTransport
+import io.libp2p.transport.implementation.PlainNettyTransport
 import io.netty.channel.ChannelHandler
-import java.net.Inet4Address
-import java.net.Inet6Address
 import java.net.InetSocketAddress
+import java.net.SocketAddress
 
 /**
  * The TCP transport can establish libp2p connections via TCP endpoints.
@@ -24,7 +21,7 @@ import java.net.InetSocketAddress
  */
 open class TcpTransport(
     upgrader: ConnectionUpgrader
-) : NettyTransport(upgrader) {
+) : PlainNettyTransport(upgrader) {
 
     override fun handles(addr: Multiaddr) =
         handlesHost(addr) &&
@@ -43,14 +40,6 @@ open class TcpTransport(
         addr: Multiaddr
     ): ChannelHandler? = null
 
-    override fun toMultiaddr(addr: InetSocketAddress): Multiaddr {
-        val proto = when (addr.address) {
-            is Inet4Address -> IP4
-            is Inet6Address -> IP6
-            else -> throw InternalErrorException("Unknown address type $addr")
-        }
-        return Multiaddr.empty()
-            .withComponent(proto, addr.address.hostAddress)
-            .withComponent(TCP, addr.port.toString())
-    } // toMultiaddr
+    override fun toMultiaddr(addr: SocketAddress): Multiaddr =
+        MultiaddrUtils.inetSocketAddressToTcpMultiaddr(addr as InetSocketAddress)
 } // class TcpTransport
